Add files via upload
Currently only have the sign up and login set up so you can test it the default insurance id is 1234, Which you can put into the insurance table before test with signing up because it needs to have a valid insurance id to sign up so it will be our testing id for now. It should allow you to see the static patient dashboard but am working on making it display the data of the actual patient. Then the logout button should work and just take you to a dummy logout page.